move the stop-serving-on-a-release to the generic apiserver
