+++ staging/src/k8s.io/apiserver/pkg/server/deleted_kinds_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controlplane
+package server
 
 import (
 	"reflect"
